% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmap_es}
\alias{cmap_es}
\title{Effect size values for Connectivity Map build 02 drugs.}
\format{A matrix where columns correspond to drugs and rows to gene symbols.}
\usage{
data(cmap_es)
}
\description{
Effect size values for all 1309 drugs in the Connectivity Map build 02.
}
\keyword{datasets}

