% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmap_tables}
\alias{cmap_tables}
\title{Top tables for Connectivity Map build 02 drugs.}
\format{A named list containing data.frames with dprime and adjusted p-values
   for each drug in CMap build 02.}
\usage{
data(cmap_tables)
}
\description{
Adjusted p-values from call to limma \code{topTables} for all 1309 Connectivity
Map build 02 drugs. Also includes unbiased effect sizes ('dprimes') from
results of call to metaMA \code{effectsize}.
}
\keyword{datasets}

