% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combo_model}
\alias{combo_model}
\title{XGBoost model for treatment combinations.}
\usage{
data(combo_model)
}
\description{
XGBoost model used to predict gene expression changes resulting from a
combination treatment.
}
\details{
Predictions for combinations are made using data from the individual
treatments. Model was trained using 'combo_train' data and is used to predict
Connectivity Map build 02 combinations using 'cmap_tables'.
}
\keyword{datasets}

