% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{HGNC symbols used for NNet predictions.}
\format{An object of class \code{character} of length 11525.}
\usage{
data(genes)
}
\value{
A character vector of 11525 HGNC symbols.
}
\description{
Order is as required for input and produced by output of net1/net2 predictions.
}
\keyword{datasets}

