% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l1000_es}
\alias{l1000_es}
\title{Effect size values for LINCS l1000 signatures.}
\format{An object of class \code{matrix} with 1001 rows and 230829 columns.}
\usage{
data(l1000_es)
}
\value{
A matrix where columns correspond to perturbagens and rows to gene symbols.
}
\description{
Moderated unbiased effect sizes values for all 230829 LINCS l1000 signatures.
}
\details{
Calculated as described by Marot et al (see reference) using \code{\link[limma]{toptable}}
from limma and \code{\link[metaMA]{effectsize}} from metaMA.
}
\references{
Marot G, Foulley JL, Mayer CD, Jaffrézic F. Moderated effect size and P-value
   combinations for microarray meta-analyses. Bioinformatics. 2009 Oct
   15;25(20):2692-9. doi: 10.1093/bioinformatics/btp444.
}
\keyword{datasets}
