% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net1}
\alias{net1}
\title{Neural network model 1 for treatment combinations.}
\format{An object of class \code{list} of length 4.}
\usage{
#NA
}
\value{
List with matrices W1/W2 and vectors b1/b2.
}
\description{
Contains weight matrices and bias vectors needed to make predictions.
}
\keyword{datasets}

