% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xgb_mod}
\alias{xgb_mod}
\title{XGBoost model for treatment combinations.}
\format{An object of class \code{xgb.Booster} of length 2.}
\usage{
#NA
}
\value{
Object of class xgb.Booster
}
\description{
Model stacks predictions from net1 and net2 with effect size values from
cmap_es and variance values from cmap_var.
}
\keyword{datasets}

