% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{l1000_es}
\alias{l1000_es}
\title{Effect size values for LINCS l1000 signatures.}
\format{An object of class \code{matrix} with 1001 rows and 230829 columns.}
\usage{
data(l1000_es)
}
\value{
A matrix where columns correspond to perturbagens and rows to gene symbols.
}
\description{
Unbiased effect sizes values for all 230829 LINCS l1000 signatures.
}
\keyword{datasets}

